add virtual liquid pot temp to pipeline

+++ genesis/utils/pipeline/data_sources/uclales.py
@@ -16,6 +16,7 @@ FIELD_NAME_MAPPING = dict(
     theta_l='t',
     cvrxp='cvrxp',
     p='p',
+    theta_l_v='theta_l_v',
 )
 
 FIELD_DESCRIPTIONS = dict(
@@ -35,6 +36,7 @@ UNITS_FORMAT = {
 
 DERIVED_FIELDS = dict(
     T=('qc', 'theta_l', 'p'),
+    theta_l_v=('theta_l', 'qv', 'qc', 'qr')
 )
 
 FN_FORMAT_3D = "3d_blocks/full_domain/{experiment_name}.tn{timestep}.{field_name}.nc"
@@ -84,14 +86,17 @@ def extract_field_to_filename(dataset_meta, path_out, field_name, **kwargs):
         field_name=field_name_src, **dataset_meta
     )
 
+    can_symlink = True
+
     if field_name_src == 'w_zt':
         path_in = path_in.parent/path_in.name.replace('.w_zt.', '.w.')
         da_w_orig = xr.open_dataarray(path_in, decode_times=False)
         da = z_center_field(da_w_orig)
+    if field_name == 'theta_l_v':
+        da = _calc_theta_l_v(**kwargs)
+        can_symlink = False
     else:
         da = xr.open_dataarray(path_in, decode_times=False)
-
-    can_symlink = True
 
     for c in "xt yt zt".split(" "):
         if 'fixes' in dataset_meta:
@@ -104,58 +109,26 @@ def extract_field_to_filename(dataset_meta, path_out, field_name, **kwargs):
 
     # the CF conventions stipulate that the long name attribute should be named
     # `long_name` not `longname`
-    da.attrs['long_name'] = da.longname
+    if 'longname' in da.attrs and not 'long_name' in da.attrs:
+        da.attrs['long_name'] = da.longname
+        can_symlink = False
 
     if field_name_src != field_name:
         can_symlink = False
         da.name = field_name
 
-    if can_symlink:
+    if can_symlink and path_in.exists():
         os.symlink(str(path_in), str(path_out))
     else:
         da.to_netcdf(path_out)
 
-    # if field_name == 'theta_v':
-        # assert 'qv' in kwargs and 't' in kwargs
 
-        # da_theta = kwargs['t'].open()
-        # da_qv = kwargs['qv'].open()
-
-        # da = _calculate_theta_v(theta=da_theta, qv=da_qv)
-        # da.name = 'theta_v'
-        # da.attrs['units'] = 'K'
-        # da.attrs['long_name'] = 'virtual potential temperature'
-    # else:
-        # ds = xr.open_dataset(path_in)
-
-
-        # da = ds[field_name_src]
-        # if field_name == 'w_zt':
-            # field_name = 'w'
-        # da.name = field_name
-
-        # new_coords = "xt yt".split(" ")
-        # old_coords = [_get_uclales_field(c) for c in new_coords]
-        # coord_map = dict(zip(old_coords, new_coords))
-        # da = da.rename(coord_map)
-        # da.attrs['long_name'] = _get_uclales_field_description(field_name)
-        # da.attrs['units'] = _cleanup_units(da.units)
-
-        # # all height levels should be the same, we do a quick check and then
-        # # use those values
-        # z__min = ds.VLEV.min(dim=old_coords)
-        # z__max = ds.VLEV.max(dim=old_coords)
-        # assert np.all(z__max - z__min < 1.0e-10)
-        # da['zt'] = np.round(z__min, decimals=3)
-        # da.zt.attrs['units'] = ds[old_coords[0]].units
-
-        # da = da.swap_dims(dict(vertical_levels='zt'))
-
-        # for c in "xt yt zt".split(" "):
-            # da.coords[c] = _scale_field(da[c])
-
-        # _scale_field(da)
-
-        # da = da.squeeze()
-
-    # da.to_netcdf(path_out)
+def _calc_theta_l_v(theta_l, qv, qc, qr):
+    # qc here refers to q "cloud", the total condensate is qc+qr (here
+    # forgetting about ice...)
+    eps = 0.608
+    theta_l_v = theta_l*(1.0 + eps*qv - (qc+qr))
+    theta_l_v.attrs['units'] = 'K'
+    theta_l_v.attrs['long_name'] = 'virtual liquid potential temperature'
+    theta_l_v.name = "theta_l_v"
+    return theta_l_v
